requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-guardian==2.4.0
django-import-export==3.3.7
django-ninja==1.1.0
django-ninja-crud==0.4.1
django-ninja-extra==0.20.4
django-ninja-jwt==5.3.0
django-simple-history==3.5.0
et-xmlfile==1.1.0
injector==0.21.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.10.0
xlrd==2.0.1
xlwt==1.3.0
